ResourceRoutesCOntroller moved to src\Http\Controllers directory
